SignOut Database

+++ lib/Objects/Account.dart
@@ -5,13 +5,13 @@ class Account
   Map<String,Account> following;
   Map<String,Account> followers;
   List<Textbook> booksOnSale;
-  double rating;
+  int rating;
   String email;
 
   Account()
   {
   }
-  Account.instantiate(String name, Map<String,Account> following, Map<String,Account> followers, List<Textbook> booksOnSale, double rating, String email)
+  Account.instantiate(String name, Map<String,Account> following, Map<String,Account> followers, List<Textbook> booksOnSale, int rating, String email)
   {
     this.name = name;
     this.following = following;
